consolidated models files into one as they all can be displayed on the user page. Figured it would be redundant to have multiple calls of User for each attraction type
